Create Music_251203a.pde
Made it so when you placed you cursor at a specific part of the screen the color would change, also depending on where you clicked a song would play. Tried to make it so every tried you left the area of a certain color when you came back the color would change but could not figure it out.